KP-7421 Remove padding on the sides of the article body
The default GitHub pages theme used with the default Jekyll build
pipeline at the time of writing uses a 980 px wide div for the page
content, whereas the pandoc-goodies version added some extra padding.
This made the text etc on our videos harder to see, so the extra padding
was removed. Padding on top and bottom was kept as is.

The "real" change was done to the _github-markdown.scss file: the rest
are autogenerated based on it.